Modify the way the process memory is scanned.
Now there are two threads.
The first thread scans the entire process memory slowly (with long delays
so it doesn't take up too much )
to look for relevant chunks.
The second thread continuously scans the relevant (whitelisted) chunks found by
the first thread.

This results in faster and more consistent response time.

In addition, scanning for the flash client is now handled slightly
differently. We now scan all Flash Client processes instead of only a
single one. This alleviates the need for the user to select the "correct"
process and seems to also fix issues with messages in the flash client not
being picked up by Tibialyzer.