Split the MainForm tabs into different files.
Every tab page is now a separate form control, the MainForm steals the
controls from these forms when it loads and creates the tab pages.

This cleans up the MainForm class significantly because all the
eventhandlers for each of the controls in the tabs are now in separate
files rather than all bunched up in the MainForm class.