Added route@ command to Tibialyzer.
This command displays a route from the players' current position to the
target position. "Route" can be chosen from the NPC, Hunting Place and
Task notifications to get a route to the specified location. In addition,
the route@ command can be used with a custom location.

Modified pathfinding algorithm to also function across
stairs/teleports/boats so pathfinding from any position to any other
position works properly.

Also added "Doors" to pathfinding (which are marked as unwalkable terrain
on the minimap, but can be walked through, so an exception has to be made
for them).

Note that not all doors and teleports are added (although a lot of them
are). In case doors/teleports are missing, the pathfinding will not
function properly.

This commit also updates nodes.db to include this new pathfinding information (including teleports and doors).