Added healing@ and damagetaken@ commands.
Changed the way damage dealt is stored, and tracked healing done in a
similar fashion as damage. Now both the "source" and "target" of both
damage and healing are stored.

On the damage@/healing@/damagetaken@ chart view, you can also filter by
the target (or source, in case of damagetaken@). So you can see ex. "who dealt
the most damage to Wyverns", or "who healed Amel Cyrom the most".